# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Personal Finance & Accounting App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: getAccounts
      tags: [accounts]
      summary: Get all user accounts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: object
                required: [accounts]
                properties:
                  accounts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create a new account
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get account by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Account details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    put:
      operationId: updateAccount
      tags: [accounts]
      summary: Update account
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountRequest"
      responses:
        "200":
          description: Account updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete account
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /accounts/{id}/summary:
    get:
      operationId: getAccountSummary
      tags: [accounts]
      summary: Get account summary with balance history
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Account summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountSummary"

  # ─── TRANSACTIONS ────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    SuccessResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
...
    Account:
      type: object
      required: [id, name, type, currencyCode, balance, isActive, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [cash, bank, card, savings, investment, other]
        currencyCode:
          type: string
        balance:
          type: number
        color:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        bankName:
          type: string
          nullable: true
        accountNumber:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    AccountSummary:
      type: object
      required: [account, transactions, balanceHistory]
      properties:
        account:
          $ref: "#/components/schemas/Account"
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        balanceHistory:
          type: array
          items:
            type: object
            required: [date, balance]
            properties:
              date:
                type: string
              balance:
                type: number
    CreateAccountRequest:
      type: object
      required: [name, type, currencyCode, balance]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [cash, bank, card, savings, investment, other]
        currencyCode:
          type: string
        balance:
          type: number
        color:
          type: string
        icon:
          type: string
        bankName:
          type: string
        accountNumber:
          type: string
    UpdateAccountRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [cash, bank, card, savings, investment, other]
        currencyCode:
          type: string
        color:
          type: string
        icon:
          type: string
        bankName:
          type: string
        accountNumber:
          type: string
        isActive:
          type: boolean

    # ─── TRANSACTIONS ─────────────────────────────────────────────
    Transaction:
      type: object
      required: [id, type, amount, currencyCode, accountId, categoryId, date, createdAt]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [income, expense, transfer]
        amount:
          type: number
        currencyCode:
          type: string
        exchangeRate:
          type: number
          nullable: true
        accountId:
          type: string
        account:
          $ref: "#/components/schemas/Account"
          nullable: true
        categoryId:
          type: string
        category:
          $ref: "#/components/schemas/Category"
          nullable: true
        description:
          type: string
          nullable: true
        receiptUrl:
          type: string
          nullable: true
        transferToAccountId:
          type: string
          nullable: true
        date:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    Category:
      type: object
      required: [id, name, type, icon, color, isDefault]
      properties:
        id:
          type: string
        name:
          type: string
        nameAr:
          type: string
          nullable: true
        type:
          type: string
          enum: [income, expense, both]
        icon:
          type: string
        color:
          type: string
        isDefault:
          type: boolean
        parentId:
          type: string
          nullable: true
...